Move EmtpyResource components to components/common
