--- airbyte-webapp/src/components/EmptyResourceBlock/index.tsx
@@ -1,3 +1,0 @@
-import EmptyResourceBlock from "./EmptyResourceBlock";
-
-export default EmptyResourceBlock;
