+++ airbyte-webapp/src/components/common/EmptyResourceBlock/index.ts
@@ -1,0 +1,1 @@
+export * from "./EmptyResourceBlock";
