+++ airbyte-webapp/src/components/common/EmptyResourceListView/EmptyResourceListView.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/colors";
+@use "scss/colors";
 
 .container {
   display: flex;
