+++ airbyte-webapp/src/pages/ConnectionPage/pages/AllConnectionsPage/AllConnectionsPage.tsx
@@ -5,7 +5,7 @@ import { FormattedMessage, useIntl } from "react-intl";
 import { useNavigate } from "react-router-dom";
 
 import { LoadingPage, MainPageWithScroll } from "components";
-import { EmptyResourceListView } from "components/EmptyResourceListView";
+import { EmptyResourceListView } from "components/common/EmptyResourceListView";
 import HeadTitle from "components/HeadTitle";
 import { Button } from "components/ui/Button";
 import { PageHeader } from "components/ui/PageHeader";
